requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
sqlmodel==0.0.34
bcrypt>=4.0
python-dotenv==1.0.1
itsdangerous==2.2.0
